bughunt/placeholders: support placeholder expansion even if '*' is in the path, add debug statements, move pattern test to own testfile
